+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added commands and support for linstor-controller token authentication
 - Added ", Corrupt Crypt Key" Warning to "resource list" and "volume list" if needed
 - Added --truncate flag to shrink table columns proportionally and truncate cells with an ellipsis when output exceeds terminal width; can also be enabled via the [global] section of the client config
+- Added --drbd-client (with aliases "--drbd-diskless-client", "-c", "--client") to
+  - resource create
+  - resource toggle-disk
+  - resource modify        [new]
 
 ### Changed